perform massive rename exercise
