+++ accelerator/models/__init__.py
@@ -316,7 +316,7 @@ from .file_page import FilePage
 from .category_header_page import CategoryHeaderPage
 from .core_profile import CoreProfile
 from .node_published_for import NodePublishedFor
-from .sidenav_association import NodeSubNavAssociation
+from .subnav_association import NodeSubNavAssociation
 from .site_redirect_page import SiteRedirectPage
 from .legal_check import LegalCheck
 from .user_legal_check import UserLegalCheck
